+++ packages/theme/src/components/index.ts
@@ -7,6 +7,8 @@ import Button from "./button"
 import Checkbox from "./checkbox"
 import CloseButton from "./close-button"
 import Code from "./code"
+import Container from "./container"
+import Divider from "./divider"
 import Drawer from "./drawer"
 import Editable from "./editable"
 import Form from "./form"
@@ -15,6 +17,7 @@ import Heading from "./heading"
 import Input from "./input"
 import Kbd from "./kbd"
 import Link from "./link"
+import List from "./list"
 import Menu from "./menu"
 import Modal from "./modal"
 import NumberInput from "./number-input"
@@ -45,6 +48,8 @@ export default {
   Checkbox,
   CloseButton,
   Code,
+  Container,
+  Divider,
   Drawer,
   Editable,
   Form,
@@ -53,6 +58,7 @@ export default {
   Input,
   Kbd,
   Link,
+  List,
   Menu,
   Modal,
   NumberInput,