+++ packages/theme/src/components/tabs.ts
@@ -1,8 +1,15 @@
 import { getColor, mode } from "@chakra-ui/theme-tools"
 
-const parts = ["tablist", "tab", "tabpanel", "indicator"]
+const parts = ["root", "tablist", "tab", "tabpanel", "indicator"]
 
 type Dict = Record<string, any>
+
+function baseStyleRoot(props: Dict) {
+  const { orientation } = props
+  return {
+    display: orientation === "vertical" ? "flex" : "block",
+  }
+}
 
 function baseStyleTab(props: Dict) {
   const { isFitted } = props
@@ -36,6 +43,7 @@ const baseStyleTabpanel = { p: 4 }
 
 const baseStyle = (props: Dict) => {
   return {
+    root: baseStyleRoot(props),
     tab: baseStyleTab(props),
     tablist: baseStyleTablist(props),
     tabpanel: baseStyleTabpanel,
@@ -68,7 +76,9 @@ const sizes = {
 
 function variantLine(props: Dict) {
   const { colorScheme: c, orientation } = props
+  const isVertical = orientation === "vertical"
   const borderProp = orientation === "vertical" ? "borderLeft" : "borderBottom"
+  const marginProp = isVertical ? "ml" : "mb"
 
   return {
     tablist: {
@@ -78,7 +88,7 @@ function variantLine(props: Dict) {
     tab: {
       [borderProp]: "2px solid",
       borderColor: "transparent",
-      mb: "-2px",
+      [marginProp]: "-2px",
       _selected: {
         color: mode(`${c}.600`, `${c}.300`)(props),
         borderColor: "currentColor",